+++ packages/swift/Sources/RustBridge/xberg-swift.swift
@@ -25604,8 +25604,8 @@ public class TesseractConfig: TesseractConfigRefMut {
     }
 }
 extension TesseractConfig {
-    public convenience init<GenericIntoRustString: IntoRustString>(_ language: RustVec<GenericIntoRustString>, _ psm: Optional<Int32>, _ output_format: GenericIntoRustString, _ oem: Int32, _ min_confidence: Double, _ preprocessing: Optional<ImagePreprocessingConfig>, _ enable_table_detection: Bool, _ table_min_confidence: Double, _ table_column_threshold: Int32, _ table_row_threshold_ratio: Double, _ use_cache: Bool, _ classify_use_pre_adapted_templates: Bool, _ language_model_ngram_on: Bool, _ tessedit_dont_blkrej_good_wds: Bool, _ tessedit_dont_rowrej_good_wds: Bool, _ tessedit_enable_dict_correction: Bool, _ tessedit_char_whitelist: GenericIntoRustString, _ tessedit_char_blacklist: GenericIntoRustString, _ tessedit_use_primary_params_model: Bool, _ textord_space_size_is_variable: Bool, _ thresholding_method: Bool) {
-        self.init(ptr: __swift_bridge__$TesseractConfig$new({ let val = language; val.isOwned = false; return val.ptr }(), psm.intoFfiRepr(), { let rustString = output_format.intoRustString(); rustString.isOwned = false; return rustString.ptr }(), oem, min_confidence, { if let val = preprocessing { val.isOwned = false; return val.ptr } else { return nil } }(), enable_table_detection, table_min_confidence, table_column_threshold, table_row_threshold_ratio, use_cache, classify_use_pre_adapted_templates, language_model_ngram_on, tessedit_dont_blkrej_good_wds, tessedit_dont_rowrej_good_wds, tessedit_enable_dict_correction, { let rustString = tessedit_char_whitelist.intoRustString(); rustString.isOwned = false; return rustString.ptr }(), { let rustString = tessedit_char_blacklist.intoRustString(); rustString.isOwned = false; return rustString.ptr }(), tessedit_use_primary_params_model, textord_space_size_is_variable, thresholding_method))
+    public convenience init<GenericIntoRustString: IntoRustString>(_ language: RustVec<GenericIntoRustString>, _ psm: Optional<Int32>, _ output_format: GenericIntoRustString, _ oem: Int32, _ min_confidence: Double, _ preprocessing: Optional<ImagePreprocessingConfig>, _ enable_table_detection: Bool, _ table_min_confidence: Double, _ table_column_threshold: Int32, _ table_row_threshold_ratio: Double, _ use_cache: Bool, _ classify_use_pre_adapted_templates: Bool, _ language_model_ngram_on: Bool, _ tessedit_dont_blkrej_good_wds: Bool, _ tessedit_dont_rowrej_good_wds: Bool, _ tessedit_enable_dict_correction: Bool, _ tessedit_char_whitelist: GenericIntoRustString, _ tessedit_char_blacklist: GenericIntoRustString, _ tessedit_use_primary_params_model: Bool, _ textord_space_size_is_variable: Bool, _ thresholding_method: GenericIntoRustString) {
+        self.init(ptr: __swift_bridge__$TesseractConfig$new({ let val = language; val.isOwned = false; return val.ptr }(), psm.intoFfiRepr(), { let rustString = output_format.intoRustString(); rustString.isOwned = false; return rustString.ptr }(), oem, min_confidence, { if let val = preprocessing { val.isOwned = false; return val.ptr } else { return nil } }(), enable_table_detection, table_min_confidence, table_column_threshold, table_row_threshold_ratio, use_cache, classify_use_pre_adapted_templates, language_model_ngram_on, tessedit_dont_blkrej_good_wds, tessedit_dont_rowrej_good_wds, tessedit_enable_dict_correction, { let rustString = tessedit_char_whitelist.intoRustString(); rustString.isOwned = false; return rustString.ptr }(), { let rustString = tessedit_char_blacklist.intoRustString(); rustString.isOwned = false; return rustString.ptr }(), tessedit_use_primary_params_model, textord_space_size_is_variable, { let rustString = thresholding_method.intoRustString(); rustString.isOwned = false; return rustString.ptr }()))
     }
 }
 public class TesseractConfigRefMut: TesseractConfigRef {
@@ -25701,8 +25701,8 @@ extension TesseractConfigRef {
         __swift_bridge__$TesseractConfig$textord_space_size_is_variable(ptr)
     }
 
-    public func thresholdingMethod() -> Bool {
-        __swift_bridge__$TesseractConfig$thresholding_method(ptr)
+    public func thresholdingMethod() -> RustString {
+        RustString(ptr: __swift_bridge__$TesseractConfig$thresholding_method(ptr))
     }
 }
 extension TesseractConfig: Vectorizable {